PR Fixes
split instrumentation to be only small runtime dispatcher; node specific code moved to instrumentation.node.ts

Logs now use ComponentState instead of just state, better consistency
